MdePkg Acpi51.h:Correct the declaration of EFI_ACPI_5_1_GTDT_SBSA_GENERIC_WATCHDOG_STRUCTURE.
Contributed-under: TianoCore Contribution Agreement 1.0
Reviewed by: Jiewen Yao <jiewen.yao@intel.com>

git-svn-id: https://svn.code.sf.net/p/edk2/code/trunk/edk2@15961 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdePkg/Include/IndustryStandard/Acpi51.h
@@ -1315,8 +1315,8 @@ typedef struct {
 ///
 typedef struct {
   UINT8   Type;
-  UINT8   Length;
-  UINT16  Reserved;
+  UINT16  Length;
+  UINT8   Reserved;
   UINT64  RefreshFramePhysicalAddress;
   UINT64  WatchdogControlFramePhysicalAddress;
   UINT32  WatchdogTimerGSIV;
